feat: refactor credential encryption and decryption logic to use dedicated module for synchronizer and ingestion credentials
