+++ opencti-platform/opencti-graphql/src/modules/ingestion/ingestion-csv-domain.ts
@@ -37,7 +37,7 @@ import { convertRepresentationsIds } from '../internal/mapper-utils';
 import { SYSTEM_USER } from '../../utils/access';
 import { regenerateCsvMapperUUID } from './ingestion-converter';
 import { createOnTheFlyUser } from '../user/user-domain';
-import { encryptIngestionCredential, decryptIngestionCredential } from '../../utils/platformCrypto';
+import { encryptIngestionCredential, decryptIngestionCredential } from './ingestion-common';
 import { findDefaultIngestionGroups } from '../../domain/group';
 
 const MINIMAL_CSV_FEED_COMPATIBLE_VERSION = '6.6.0';
